refactor(Example NextJS): Remove Contentlets component and introduce RecommendedCard, EditButton, and ReorderMenuButton for enhanced modularity
